+++ pypsa/optimization/optimize.py
@@ -19,7 +19,6 @@ from linopy.solvers import available_solvers
 
 from pypsa._options import options
 from pypsa.common import UnexpectedError, as_index
-from pypsa.components._types.mixin.multiports import _Multiport
 from pypsa.components.array import _from_xarray
 from pypsa.components.common import as_components
 from pypsa.consistency import check_big_m_exceeded, check_no_modular_committables
@@ -372,7 +371,7 @@ def define_objective(
             cost_weight = c.da.active.sel(name=ext_i).any(dim="snapshot")
 
         y_attr = "capital_cost"
-        pw_attr = c._piecewise_schema(y=y_attr)
+        pw_attr = c._piecewise_schema(y_attr)
         if not pw_attr.empty:
             x_var = m[f"{c.name}-{pw_attr.x}"]
             extra_options = filter(
@@ -1051,15 +1050,13 @@ class OptimizationAccessor(OptimizationAbstractMixin):
                 )
                 continue
             c = n.c[_c_name]
-            pw_schema = c._piecewise_schema(aux_variable=attr)
+            rows = c._piecewise_attrs
+            pw_schema = rows[rows.aux_variable == attr].squeeze()
+            if pw_schema.empty and attr.endswith("_piecewise"):
+                # We deal with these variables below when processing the `p` attr.
+                continue
             # Piecewise variables are auxiliary and need to be processed before being passed back as a solution.
             if not pw_schema.empty:
-                if (
-                    isinstance(c, _Multiport)
-                    and c._get_base_coeff(pw_schema.y) == c._coefficient_attr
-                ):
-                    # We deal with these variables below when processing the `p` attr.
-                    continue
                 x_var = m.variables[c._piecewise_x_var(pw_schema.y)]
                 sol = sol / x_var.solution
                 if "snapshot" in sol.dims: